Split pg tests into connector tests and store tests
