+++ tests/integration/test_pg_store.py
@@ -1,15 +1,28 @@
 import datetime
-import functools
-import json
 import random
 import string
 
 import pendulum
 import pytest
 
-from procrastinate import aiopg_connector, jobs
+from procrastinate import jobs, store
 
 pytestmark = pytest.mark.asyncio
+
+
+@pytest.fixture
+def pg_job_store(pg_connector):
+    return store.JobStore(connector=pg_connector)
+
+
+@pytest.fixture
+def get_all(pg_connector):
+    async def f(table, *fields):
+        return await pg_connector.execute_query_all(
+            f"SELECT {', '.join(fields)} FROM {table}"
+        )
+
+    return f
 
 
 @pytest.mark.parametrize(
@@ -100,7 +113,7 @@ async def test_get_job_no_result(pg_job_store, job):
     assert await pg_job_store.fetch_job(queues=["queue_a"]) is None
 
 
-async def test_get_stalled_jobs(get_all, pg_job_store):
+async def test_get_stalled_jobs(get_all, pg_job_store, pg_connector):
     await pg_job_store.defer_job(
         jobs.Job(
             id=0,
@@ -117,7 +130,7 @@ async def test_get_stalled_jobs(get_all, pg_job_store):
 
     # We start a job and fake its `started_at`
     job = await pg_job_store.fetch_job(queues=["queue_a"])
-    await pg_job_store.execute_query(
+    await pg_connector.execute_query(
         f"UPDATE procrastinate_jobs SET started_at=NOW() - INTERVAL '30 minutes' "
         f"WHERE id={job_id}"
     )
@@ -140,7 +153,7 @@ async def test_get_stalled_jobs(get_all, pg_job_store):
     )
 
 
-async def test_delete_old_jobs_job_is_not_finished(get_all, pg_job_store):
+async def test_delete_old_jobs_job_is_not_finished(get_all, pg_job_store, pg_connector):
     await pg_job_store.defer_job(
         jobs.Job(
             id=0,
@@ -158,7 +171,7 @@ async def test_delete_old_jobs_job_is_not_finished(get_all, pg_job_store):
     # We start a job
     job = await pg_job_store.fetch_job(queues=["queue_a"])
     # We back date the started event
-    await pg_job_store.execute_query(
+    await pg_connector.execute_query(
         f"UPDATE procrastinate_events SET at=at - INTERVAL '2 hours'"
         f"WHERE job_id={job.id}"
     )
@@ -168,7 +181,7 @@ async def test_delete_old_jobs_job_is_not_finished(get_all, pg_job_store):
     assert len(await get_all("procrastinate_jobs", "id")) == 1
 
 
-async def test_delete_old_jobs_multiple_jobs(get_all, pg_job_store):
+async def test_delete_old_jobs_multiple_jobs(get_all, pg_job_store, pg_connector):
     await pg_job_store.defer_job(
         jobs.Job(
             id=0,
@@ -195,7 +208,7 @@ async def test_delete_old_jobs_multiple_jobs(get_all, pg_job_store):
     await pg_job_store.finish_job(job_a, status=jobs.Status.SUCCEEDED)
     await pg_job_store.finish_job(job_b, status=jobs.Status.SUCCEEDED)
     # We back date the events for job_a
-    await pg_job_store.execute_query(
+    await pg_connector.execute_query(
         f"UPDATE procrastinate_events SET at=at - INTERVAL '2 hours'"
         f"WHERE job_id={job_a.id}"
     )
@@ -207,7 +220,7 @@ async def test_delete_old_jobs_multiple_jobs(get_all, pg_job_store):
     assert rows[0]["id"] == job_b.id
 
 
-async def test_delete_old_job_filter_on_end_date(get_all, pg_job_store):
+async def test_delete_old_job_filter_on_end_date(get_all, pg_job_store, pg_connector):
     await pg_job_store.defer_job(
         jobs.Job(
             id=0,
@@ -222,7 +235,7 @@ async def test_delete_old_job_filter_on_end_date(get_all, pg_job_store):
     # We finish the job
     await pg_job_store.finish_job(job, status=jobs.Status.SUCCEEDED)
     # We back date only the start event
-    await pg_job_store.execute_query(
+    await pg_connector.execute_query(
         f"UPDATE procrastinate_events SET at=at - INTERVAL '2 hours'"
         f"WHERE job_id={job.id} AND TYPE='started'"
     )
@@ -250,7 +263,14 @@ async def test_delete_old_job_filter_on_end_date(get_all, pg_job_store):
     ],
 )
 async def test_delete_old_jobs_parameters(
-    get_all, pg_job_store, status, nb_hours, queue, include_error, should_delete
+    get_all,
+    pg_job_store,
+    pg_connector,
+    status,
+    nb_hours,
+    queue,
+    include_error,
+    should_delete,
 ):
     await pg_job_store.defer_job(
         jobs.Job(
@@ -267,7 +287,7 @@ async def test_delete_old_jobs_parameters(
     # We finish the job
     await pg_job_store.finish_job(job, status=status)
     # We back date its events
-    await pg_job_store.execute_query(
+    await pg_connector.execute_query(
         f"UPDATE procrastinate_events SET at=at - INTERVAL '2 hours'"
         f"WHERE job_id={job.id}"
     )
@@ -327,12 +347,13 @@ async def test_finish_job_retry(get_all, pg_job_store):
     assert job2.attempts == job1.attempts + 1
 
 
-async def test_listen_queue(pg_job_store):
+async def test_listen_for_jobs(pg_job_store, pg_connector):
     queue = "".join(random.choices(string.ascii_letters, k=10))
     queue_full_name = f"procrastinate_queue#{queue}"
+
     await pg_job_store.listen_for_jobs(queues=[queue])
 
-    count = await pg_job_store.execute_query_one(
+    count = await pg_connector.execute_query_one(
         """SELECT COUNT(*) FROM pg_listening_channels()
                           WHERE pg_listening_channels = %(queue)s""",
         queue=queue_full_name,
@@ -340,20 +361,20 @@ async def test_listen_queue(pg_job_store):
     assert count["count"] == 1
 
 
-async def test_listen_all_queue(pg_job_store):
+async def test_listen_for_jobs_all_queue(pg_job_store, pg_connector):
     await pg_job_store.listen_for_jobs()
 
-    count = await pg_job_store.execute_query_one(
+    count = await pg_connector.execute_query_one(
         """SELECT COUNT(*) FROM pg_listening_channels()
            WHERE pg_listening_channels = 'procrastinate_any_queue'"""
     )
     assert count["count"] == 1
 
 
-async def test_enum_synced(pg_job_store):
+async def test_enum_synced(pg_connector):
     # If this test breaks, it means you've changed either the task_status PG enum
     # or the python procrastinate.jobs.Status Enum without updating the other.
-    pg_enum_rows = await pg_job_store.execute_query_all(
+    pg_enum_rows = await pg_connector.execute_query_all(
         """SELECT e.enumlabel FROM pg_enum e
                JOIN pg_type t ON e.enumtypid = t.oid WHERE t.typname = %(type_name)s""",
         type_name="procrastinate_job_status",
@@ -362,60 +383,6 @@ async def test_enum_synced(pg_job_store):
     pg_values = {row["enumlabel"] for row in pg_enum_rows}
     python_values = {status.value for status in jobs.Status.__members__.values()}
     assert pg_values == python_values
-
-
-async def test_get_connection(connection):
-    dsn = connection.dsn
-    async with await aiopg_connector.get_connection(dsn=dsn) as new_connection:
-
-        assert new_connection.dsn == dsn
-
-
-async def test_get_connection_json_loads(connection, mocker):
-    dsn = connection.dsn
-    json_loads = mocker.MagicMock()
-    register_default_jsonb = mocker.patch("psycopg2.extras.register_default_jsonb")
-    async with await aiopg_connector.get_connection(
-        dsn=dsn, json_loads=json_loads
-    ) as new_connection:
-        register_default_jsonb.assert_called_with(new_connection.raw, loads=json_loads)
-
-
-async def test_execute_query_one_json_loads(connection, mocker):
-    class NotJSONSerializableByDefault:
-        pass
-
-    def encode(obj):
-        if isinstance(obj, NotJSONSerializableByDefault):
-            return "foo"
-        raise TypeError()
-
-    query = "SELECT %(arg)s::jsonb as json"
-    arg = {"a": "a", "b": NotJSONSerializableByDefault()}
-    json_dumps = functools.partial(json.dumps, default=encode)
-    result = await aiopg_connector.execute_query_one(
-        connection, query, json_dumps=json_dumps, arg=arg
-    )
-    assert result["json"] == {"a": "a", "b": "foo"}
-
-
-async def test_execute_query_all_json_loads(connection, mocker):
-    class NotJSONSerializableByDefault:
-        pass
-
-    def encode(obj):
-        if isinstance(obj, NotJSONSerializableByDefault):
-            return "foo"
-        raise TypeError()
-
-    query = "SELECT %(arg)s::jsonb as json"
-    arg = {"a": "a", "b": NotJSONSerializableByDefault()}
-    json_dumps = functools.partial(json.dumps, default=encode)
-    result = await aiopg_connector.execute_query_all(
-        connection, query, json_dumps=json_dumps, arg=arg
-    )
-    assert len(result) == 1
-    assert result[0]["json"] == {"a": "a", "b": "foo"}
 
 
 async def test_defer_job(pg_job_store, get_all):
@@ -437,56 +404,3 @@ async def test_defer_job(pg_job_store, get_all):
             "task_name": "bob",
         }
     ]
-
-
-async def test_execute_query(pg_job_store):
-    await pg_job_store.execute_query(
-        "COMMENT ON TABLE \"procrastinate_jobs\" IS 'foo' "
-    )
-    result = await pg_job_store.execute_query_one(
-        "SELECT obj_description('public.procrastinate_jobs'::regclass)"
-    )
-    assert result == {"obj_description": "foo"}
-
-    result = await pg_job_store.execute_query_all(
-        "SELECT obj_description('public.procrastinate_jobs'::regclass)"
-    )
-    assert result == [{"obj_description": "foo"}]
-
-
-async def test_close_connection(pg_job_store):
-    await pg_job_store.get_connection()
-    await pg_job_store.close_connection()
-    assert pg_job_store._connection.closed == 1
-
-
-async def test_close_connection_no_connection(pg_job_store):
-    await pg_job_store.close_connection()
-    # Well we didn't crash. Great.
-
-
-async def test_stop_no_connection(pg_job_store):
-    pg_job_store.stop()
-    # Well we didn't crash. Great.
-
-
-async def test_get_connection_called_twice(pg_job_store):
-    conn1 = await pg_job_store.get_connection()
-    assert not conn1.closed
-    conn2 = await pg_job_store.get_connection()
-    assert conn2 is conn1
-
-
-async def test_get_connection_after_close(pg_job_store):
-    conn1 = await pg_job_store.get_connection()
-    assert not conn1.closed
-    await pg_job_store.close_connection()
-    conn2 = await pg_job_store.get_connection()
-    assert not conn2.closed
-    assert conn2 is not conn1
-
-
-async def test_get_connection_no_psycopg2_adapter_registration(pg_job_store, mocker):
-    register_adapter = mocker.patch("psycopg2.extensions.register_adapter")
-    await pg_job_store.get_connection()
-    assert not register_adapter.called
